mining: Move TxMiningView tests to mining.
This moves the TxMiningView tests from mempool to mining, since
TxMiningView is now located in the mining package.
